Moved {agentid}/services to agents implementation.
